backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-search-results
